pyproject.toml
[project]
name = "resume-tailoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "keybert>=0.9.0",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.5",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]
